3A-2 OriginMJ.java要做很多事情...累 TODO 要做帶參數的ArrayList 讓四家只帶一個參數就能掌握自己位置 3A-3 四家發牌OK 打牌能打到海底OK 一起監聽改變數值OK
3B-5 21點修繕95%完成 基本上OK了 剩下畫面

+++ app/src/main/java/com/example/mahjongv2/SmallGameActivity.java
@@ -279,7 +279,7 @@ public class SmallGameActivity extends AppCompatActivity {
         if(f >= 11.109f && f <=11.111f){
             //BlackJack youWin
             Log.v("leo","blackJack");
-            return ""+Math.floor(f);
+            return ""+(int)Math.floor(f);
             //A + A~10
         }else if((int)(Math.floor(f*100)) - (int)(Math.floor(f*10)*10) >0 && f<12){
             return (int)Math.floor(f)+"/"+((int)Math.floor(f)+10);
